requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
mediapipe
snowflake-connector-python
python-dotenv